Result requests

+++ src/Models/Quiz.cs
@@ -1,92 +1,92 @@
 namespace AnalyseMeAPI.Models {
     public class _Compass {
-        public int econ { get; set; }
-        public int govt { get; set; }
-        public int scty { get; set; }
+        public float econ { get; set; }
+        public float govt { get; set; }
+        public float scty { get; set; }
     }
 
     public class _8Values {
-        public int econ { get; set; }
-        public int dipl { get; set; }
-        public int govt { get; set; }
-        public int scty { get; set; }
+        public float econ { get; set; }
+        public float dipl { get; set; }
+        public float govt { get; set; }
+        public float scty { get; set; }
     }
 
     public class _9Axes {
-        public int fed { get; set; }
-        public int dem { get; set; }
-        public int glo { get; set; }
-        public int mil { get; set; }
-        public int fre { get; set; }
-        public int equ { get; set; }
-        public int sec { get; set; }
-        public int pro { get; set; }
-        public int mul { get; set; }
+        public float fed { get; set; }
+        public float dem { get; set; }
+        public float glo { get; set; }
+        public float mil { get; set; }
+        public float fre { get; set; }
+        public float equ { get; set; }
+        public float sec { get; set; }
+        public float pro { get; set; }
+        public float mul { get; set; }
     }
 
     public class _Soulgraphy_PI {
-        public int econ_soci { get; set; }
-        public int econ_regu { get; set; }
-        public int econ_prot { get; set; }
-        public int state_glob { get; set; }
-        public int state_peac { get; set; }
-        public int state_fede { get; set; }
-        public int state_demo { get; set; }
-        public int society_ind { get; set; }
-        public int society_prog { get; set; }
-        public int society_secu { get; set; }
-        public int society_revo { get; set; }
-        public int law_free { get; set; }
-        public int law_reso { get; set; }
-        public int law_ecol { get; set; }
+        public float econ_soci { get; set; }
+        public float econ_regu { get; set; }
+        public float econ_prot { get; set; }
+        public float state_glob { get; set; }
+        public float state_peac { get; set; }
+        public float state_fede { get; set; }
+        public float state_demo { get; set; }
+        public float society_ind { get; set; }
+        public float society_prog { get; set; }
+        public float society_secu { get; set; }
+        public float society_revo { get; set; }
+        public float law_free { get; set; }
+        public float law_reso { get; set; }
+        public float law_ecol { get; set; }
     }
 
     public class _Soulgraphy_EFA {
-        public int org_prop { get; set; }
-        public int org_law { get; set; }
-        public int org_hier { get; set; }
-        public int tax_prog { get; set; }
-        public int tax_arti { get; set; }
-        public int tax_work { get; set; }
-        public int tax_weal { get; set; }
-        public int reg_busi { get; set; }
-        public int reg_labo { get; set; }
-        public int reg_mone { get; set; }
-        public int reg_bure { get; set; }
-        public int reg_ecol { get; set; }
-        public int market_trad { get; set; }
-        public int market_auta { get; set; }
-        public int market_inve { get; set; }
-        public int market_bank { get; set; }
+        public float org_prop { get; set; }
+        public float org_law { get; set; }
+        public float org_hier { get; set; }
+        public float tax_prog { get; set; }
+        public float tax_arti { get; set; }
+        public float tax_work { get; set; }
+        public float tax_weal { get; set; }
+        public float reg_busi { get; set; }
+        public float reg_labo { get; set; }
+        public float reg_mone { get; set; }
+        public float reg_bure { get; set; }
+        public float reg_ecol { get; set; }
+        public float market_trad { get; set; }
+        public float market_auta { get; set; }
+        public float market_inve { get; set; }
+        public float market_bank { get; set; }
     }
 
     public class _Soulgraphy_PFA {
-        public int expr_spee { get; set; }
-        public int expr_art { get; set; }
-        public int expr_poli { get; set; }
-        public int state_anar { get; set; }
-        public int state_forc { get; set; }
-        public int state_poli { get; set; }
-        public int state_huma { get; set; }
-        public int state_surv { get; set; }
-        public int priv_priv { get; set; }
-        public int priv_drug { get; set; }
-        public int priv_guns { get; set; }
-        public int priv_mora { get; set; }
-        public int priv_reli { get; set; }
-        public int eq_oppo { get; set; }
-        public int eq_resu { get; set; }
-        public int eq_burd { get; set; }
-        public int eq_free { get; set; }
+        public float expr_spee { get; set; }
+        public float expr_art { get; set; }
+        public float expr_poli { get; set; }
+        public float state_anar { get; set; }
+        public float state_forc { get; set; }
+        public float state_poli { get; set; }
+        public float state_huma { get; set; }
+        public float state_surv { get; set; }
+        public float priv_priv { get; set; }
+        public float priv_drug { get; set; }
+        public float priv_guns { get; set; }
+        public float priv_mora { get; set; }
+        public float priv_reli { get; set; }
+        public float eq_oppo { get; set; }
+        public float eq_resu { get; set; }
+        public float eq_burd { get; set; }
+        public float eq_free { get; set; }
     }
 
     public class _MyPolitics {
-        public int social { get; set; }
-        public int lesef { get; set; }
-        public int anarchy { get; set; }
-        public int peace { get; set; }
-        public int ecology { get; set; }
-        public int progress { get; set; }
-        public int glob { get; set; }
+        public float social { get; set; }
+        public float lesef { get; set; }
+        public float anarchy { get; set; }
+        public float peace { get; set; }
+        public float ecology { get; set; }
+        public float progress { get; set; }
+        public float glob { get; set; }
     }
 }
